requirements.txt
numpy
pandas
scipy
joblib
matplotlib
seaborn
matplotlib
jupyter
scikit-learn
python-dotenv
psycopg2
mlflow
python-dotenv
fastapi[standard]
